refactor(Core): Move network to a separate package
