+++ lib/core/infrastructure/network/dio_client.dart
@@ -1,4 +1,4 @@
-import 'package:currency_exchange/core/infrastructure/base_network_client.dart';
+import 'package:currency_exchange/core/infrastructure/network/base_network_client.dart';
 import 'package:currency_exchange/util/api_constants.dart';
 import 'package:currency_exchange/util/di_named_keys_constants.dart';
 import 'package:dio/dio.dart';
